// apps/web/interfaces/Restaurant.ts
export default interface Restaurant {
    readonly id: string
    readonly area: string
    readonly url: string
    readonly name: string
    readonly genre: string
    readonly tel: string
    readonly businessDayInfo: string
    readonly address: string
    readonly latitude: string
    readonly longitude: string
    readonly distance: string
    readonly isOpen: number
}